Initial commit
This is the first commit for this project. Since this project is a derivative of PDE-LEARN, I started by copying PDE-LEARN. This commit is literally just a duplicate of PDE-LEARN. In my next commit, I plan to remove the bits of PDE-LEARN that are not applicible to this project. I will then begin building up new code (starting with a weight function class, derivative class, library function class, and integral function... along with modifications to the settings reader)