Removed the bits of PDE-READ that I don't need.
I also created the files (though not their content) for the new code. I moved the reader stuff to a separate directory, called IO (the printing stuff is in here too). I decided to do the following for reading in a library:

The settings file will point to a second file. This second file will contain text that specifies the trial functions and derivatives (from which, we construct the library terms). Currently, my plan is something like the following: the file will contain a list. Each list item consists of a trial function and a derivative. Trial functions are functions of u. Together, a derivative applied to a trial function gives a library term (a term in the PDE). We assume the PDE has the following form:

D^{(0)} F_0(U) = c_1 D^{(1)} F_1(U) + ... + c_N D^{(N)} F_N(U)

The term on the right, F_0, is called the "Left-Hand side term" or "LHS Term" for short. The terms on the right are called "Right-Hand side terms" or "RHS Terms" for short. Each term is a "library term". The functions F_0, ... , F_N are the "trial functions", and the derivative operators D^{(0)}, ... , D^{(N)} are the "derivatives". The list will contain a special listing just for the LHS Term, followed by a section to list the RHS terms. This is somewhat crude (a lot less sophistiated than the "automatic library generation" of PDE-LEARN, but I think it should work. Each derivative should just be a list of partial derivatives. Something like "D_x^2 D_t". I'll make the reader parse something like that. Each library term will probably just be a power of U (something like U^3), or maybe even sinusodials. Thus, each line of the Library file will look something like the following:

D_x^2 U*Sin(U)

The reader will parse the derivative and the library term. To make things easy, I may require that the trial function part contains no spaces so that I can split by spaces (and then split the last part by *). To be determined... My plan is to then have three classes: one for derivatives (which will just be a variable length array... each array element is an integer which corresponds to a particular partial derivative. For example, D_x^2 D_t might become [1, 0]), a second for trial functions (initially, this will probably just be U to some power... I'll modify it down the road if I want to add in sinusodials or other things like that), and one for library terms (which will hold a derivative and a trial function). The idea is that you can call a library function for a particular U + coords.

I also want to write a weight function class. This will need to define the weight function, as well as its partial derivatives (these will need to be hard coded). Since we plan to use the same "flavor" of weigth function for all weight functions (namely, bumps), I think I can make a "bump function" class, with derivatives hard coded, that is paramaterized in terms of the bump's center and diameter.

This will take some time to implement. Expect the library stuff first, with weight function stuff down the road.