Added a lot of stuff...
In summary, I got most of the "Library.txt reader" running. To do this, I had to add a lot of files/code...

Derivative.py, Trial_Function.py, and Library_Term.py all implement classes. The Derivative class holds an abstract representation of a partial derivative operator. This class is not designed to actually implement a partial derivative operator. Rather, it holds information about which partial deriatives are in the derivative operator. The idea is the following: We want to compute the integral of w(t, x, y, z) D^{(k)} F_k(U(t,x, y, z)) where w is a weight function, D^{(k)} is some partial derivative operator, and F_k is a trial function. We use the magic of integration by parts to rewrite this as the integral of (-1)^{m(k)}D^{(k)} w(t, x, y, z) F_k(U(t, x, y, z)), where m(k) is the sum of the indices in D^{(k)}. Since w is known ahead of time (it's a bump function), we can calculate the derivatives of the various weight functions ahead of time. There are two ways to do this: either we can write down an analytic expression for the derivatives, then evaluate each derivative for each weight function, storing the results. Alternatively, we can use automatic differentiation to compute the derivatives ahead of time, then store the values. Either way, we only have to evaluate these derivatives once. We can then reuse the values each time we compute the integral. To compute these derivatives, however, we need a way of determining which derivatives to evaluate, this is precisely what the Derivative class is supposed to do. It holds that information.

The Trial_Function class holds a function of U (one of the F_k's in the representation of the PDE). Currently, this is just a natural number power of U. I may expand this in the future.

The Library_Term class holds a Derivative object and a Trial_Function objcect. This is designed to implement a proper library function.

The Library_Reader.py file essentially reads the contents of Library.txt, and converts those contents into objects from these classes. It then returns a tuple, whose first element contains the RHS Library term, and whose second term is a list of the LHS Library terms.

Currently, the code appears to work. In my next commit, I want to integrate this in with the settings reader. I then want to begin working on the weight function. Initially, I think my plan is the following: First, specify the quadrature points, number of weight functioins, their centers, and their radii. Next, determine the highest order partial derivative we need with respect to each variable. Suppose this is m(t) for t, m(x) for x, etc. We can then systematically evaluate all derivatives of the weight functions of the form D_t^{m(t)} D_x^{m(x)} ... D_z^{(m(z)}. To do this, we first evaluate D_t w, D_x w, D_y w, D_z w at the quadrature points. Next, we use D_t w to evaluate D_t^2 w, D_t D_x w, ... , D_t D_z w at the quadrature points. We then use D_x w to compute D_x^2 w, D_x D_y w, D_x D_z w at the quadrature points and so on. At each step, we store all of these results. We keep doing this until we get D_t^{m(t)} D_x^{m(x)} ... D_z^{m(z)} w. Next, we only keep those derivatives of w corresponding to derivatives that are present in the library terms in the PDE. It may also be reasonable (in terms of computation time to just calculate D^{(k)} w for each derivative D^{(k)} that is present in the PDE. This redoes computations... but maybe that's okay... since we only compute them once? Idk. Something like that. Either way, do this for each weight function, and store the results in the weight function object.

Once I have the weight functions + derivative calculation thing working, I plan to write an integrate function. This will take a library term and a weight function, and compute their integral. I plan to be smart about this, only integrating over points corresponding to the support of the weight function (and its derivatives). Expect that in a while.