Started Weight Function class
I started a class for weight functions. Currently, it just intializes the class/defines a forward method. I decided to make the weight function a torch.nn.Module subclass... I did this mainly because of how I plan to calculate the derivatives of a weigth function... I need torch to be able to differentiate stuff that I pass through them. To do this, I need a weight function to be a torch.nn.Module object with a forward method.

I also wrote a simple function (in the new plot directory) to plot a slice of the output of a weight function (mainly for debugging purposes). In my next commit, I plan to add support for calculating derivatives of the weight function. To make sure that things work, I plan to add plotters for the derivatives as well.