Wrote code to automatically compute w derivatives
I rewrote the "Evaluate_Derivatives" function (though forgot to give it a new docstring). It is now designed to take an arbitrary partial derivative of a weight function. Amazingly, it appears to do this... and do it well. I also wrote a plot function to plot various derivatives (mainly for debugging purposes).

In my next commit, I need to update Evaluate_Derivative's docstring. I then need to write code that takes in a list of derivative operators and generates the corresonding derivatives of w, storing them in w's dictionary. Once I have that, all I need to write is a "Integrate" function, and then the code will be mostly ready to test.