Weight function class now only evaluates derivatives in its support
I did this by making the coordinates at which we evaluate the derivatives an explicit attribute of the weight function class. This allows me to determine which coordinates are in the support of the bump function, and which are not. I only calculate the weight function and its derivatives in its support, then use some clever book-keeping to recover the bump function/its derivatives on the entire domain when I need that. This should dramatically reduce storage demands + make integration a lot faster (when I write it)