Wrote an integrate function + Tests
I wrote a function that integrates a weight function using a multi-dimensional composite trapezodial rule. This code is located in Integrate.py. I also wrote some tests for it. Currently, I don't have a good way to generate traditional unit tests. However, I was able to compare the output of my code with that of mathematica. It seems that my code works/gives reasonably accurate values for the integral (both for the weight function, and some of its derivatives). With that said, the code does appear to run into some trouble with higher order derivatives. These higher order derivatives are non-zero only on a small subset of the problem domain. This means that if the grid is not sufficiently fine, the integral will be reasonably inaccurate. With that said, I 200 or so gridlines along each axis seems to work well for the problem domain [-3, 3] x [-3, 3] x [-3, 3] witha radius of 1.

I also made minor tweaks to some other code. I am unsure if I will keep this integrate function. I just wanted something that's reasonably accurate/easy to implement, so that I can quickly get the code to the point where it can actually learn PDEs. In my next commit, I plan to write the Weak_Form_Loss function, and update main.py to generate the weight functions, apply each derivative to those weight functions, and then train the networks.