I got the code running
The main changes in this commit are in Loss (I wrote Weak From Loss) and main (I wrote code to generate weight functions with random centers in the domain, generate a grid of coordinates, and print out the resulting PDE). I had to fix a tonnnnn of bugs and typos (even in the data set generation code)... but it runs.

There appear to be some problems (it's not really learning Burgers equation...), but it is at least running :D