Added a plotter for 2D solutions
This is pretty straight forward, it just plots the output of a network, U, that it loads from from a save. The code is fairly primitive right now (you have to supply bounds and such). I may add a settings file/automate some things in the future. The network, U, appears to learn the burgers equation data set, which implies that the problem lies elsewhere (not with the data, or U).