Reorganized + updated weak form loss.
I renamed "IO" as "Readers", and moved all files that define a class into a new directory called "Classes". This cleaned up the code. I also made minor tweaks elsewhere.

I also changed the weak form loss by dividing it by the number of rows (so it's the mean square residual, not the sum of residuals... the latter is dependent on the number of weight functions, which is not a good thing). This make make the weak form loss too small, which may force me to multiply it by a constant... to be determined