Tweaked weight function
I ran into some issues with the Evaluate_Derivatives function. For reasons that I do not entirely understand, nan's appear when evaluating high order derivatives of weight functions near the edge of their support. This is odd, because those derivatives should be very close to zero. I don't really know why it's happening, because it results from using the auto-diff function. This appears to be a bug with autodiff. I say that because it doesn't happen at other points, and it disappears when swtiching to double precision, or moving a little bit away from the boundary (switching to double precision yields the small values). I'm assuming that something in auto diff's implementation causes inf - inf, or something like that. I don't really understand it. I spent a while tweaking the weight function implementation, but ultimeatly concluded that I could not prevent it by adjusting forward (the bad computations take place using a graph that torch finds... I don't know what happens in this graph, but it's bad). I did, however, find a temporary fix: cast the Coordaintes (in Evaluate_Derivative) to double, perform calculations using double precision, and then cast the result back to single precision when returning. This fixed the issue.

With this change, I was able to learn the infamous KS equation. I have yet to try anything with sparsity or noise; mainly, I just want to run a few tests right now to hunt down bugs/make tweaks.

I do not have any immediate plans for the next commit. I may try upgrading the plotter (a settings file... maybe?), or makking weight function radius a setting (currenlty, it's set to a fraction of the smallest sub-interval of the problem domain (a rectangle)... though I don't like this approach).