Updated plotter
I overhauled the "Plot_U_2D" function. First, I wrote a settings file and setting reader for it. I also shuffled a few things in the readers (moving "Read_Setting" and "Reat_Bool_Setting" to "File_Reader"). The new plot settings file requires a matlab data set. It accepts a save name, network architecture, and matlab (.mat) dataset. I then changed the main Plot_U_2D function. It now makes two plots; one depicts the learned solution, the other depicts the error between the data set and the learned solution at the corresponding inputs. I had to specify a lotttt of settings to make the plot not look terrible.. but it worked.

I also tested KS with 50% noise. It worked flawlessly; the coefficients have an error of < 4% (with 2000 epochs of Adam, 100 of LBFGS). I have not yet decided how to select the weight function radius (it's jsut hardcoded to be 1/4th of the length of the shortest side length of the problem domain). I also have not really tested the code with vairous numbers of weight functions / integrals.

I'm also considering adding a hyperparamter constant (like lambda) infront of the weak form loss. Currently, I haven't needed one, but I can forsee a future in which I do.