Rewrote integrate function
The idea is that the weight function store the coordinates. The integrate function evaluates U on these coordinates, then computes the trial functions on the coordinates, then uses these values to compute the integral. The weight function also stores the volume, V, of each sub-rectangle engendered by the grid of coordinates.

This new approach will make it much much much easier to switch to the "master weight function" idea.