Update settings, fixed some bugs
Apparently, Python was written by a couple of 7 year olds. If you include escape characters ('\') in a docstring, it freaks out and has no idea what to do. A docstring is a COMMENT, so its contents should have no impact on the code. Somehow, that basic principle flew over the heads of the miserable 7 year olds who wrote Python. As such, if you include backslash characters in a docstring, Python might throw a tantrum. They wrote in a hack of a fix, however... put an "r" infront of the docstring. I guess that fixes it... I hate when 7 year olds write programming languages.

Anyway, I also updated the settings reader + Settings file. In my next commit, I plan to rewrite the main file and add a plotter... then I can start debugging... woo-hoo. I also need to fix the plotter (to use the new settings format).